30134751_music_assessement/requirements.txt
djangorestframework
factory_boy
locust
pyJWT
python-decouple
pytest
pytest-benchmark
pytest-django
Pillow